Revert "Some small changes"
This reverts commit 69564a0b1081a8b38de96bfc7ec3c052a6dd4ba3.

+++ source/src/XT_HashTable.pas
@@ -11,9 +11,9 @@ uses
 
 type
   ColorLAB = Record
-    L: Double;
-    A: Double;
-    B: Double;
+    L: Single;
+    A: Single;
+    B: Single;
   end;
 
   ColorLABEntry = Record
